Moving complex helpers to clients/utility.
