+++ clients/include/utility.h
@@ -487,6 +487,45 @@ void print_matrix(vector<T> CPU_result, vector<T> GPU_result, int m, int n, int 
         }
 }
 
+/* =============================================================================================== */
+/* Complex / real helpers.                                                                         */
+template <typename T>
+constexpr bool is_complex = false;
+
+template <>
+constexpr bool is_complex<hipblasComplex> = true;
+
+template <>
+constexpr bool is_complex<hipblasDoubleComplex> = true;
+
+// Get base types from complex types.
+template <typename T, typename = void>
+struct real_t_impl
+{
+    using type = T;
+};
+
+template <typename T>
+struct real_t_impl<T, std::enable_if_t<is_complex<T>>>
+{
+    using type = decltype(T().x);
+};
+
+namespace std
+{
+    template <typename T>
+    T abs(const hip_complex_number<T>& z)
+    {
+        T tr = abs(z.x), ti = abs(z.y);
+        return tr > ti ? (ti /= tr, tr * sqrt(ti * ti + 1)) : (tr /= ti, ti * sqrt(tr * tr + 1));
+    }
+}
+
+template <typename T>
+using real_t = typename real_t_impl<T>::type;
+
+/* =============================================================================================== */
+
 #ifdef __cplusplus
 extern "C" {
 #endif
